PREDICTION OF PRICE OF A WINE USING REGRESSION
Wineinformatics is a field that uses machine-learning and data-mining techniques to glean useful information from wine. Wine is one of the most popular drinks in the world, with over twenty-eight billion liters produced across 63 countries in the year 2015 alone. To be this popular, wine must have several interesting characteristics which humans enjoy: aroma, color and flavor. Many computer-based techniques have been applied or developed for use in the wine field, such as software for wine-making and classification of the wines’ characteristics. Wineinformatics studies how these characteristics can be used to make inferences about the wine, such as the wine’s quality and how expensive it may be.